pyproject.toml
[project]
name = "jejuchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.9.0",
    "geopy>=2.4.1",
    "google-generativeai>=0.8.3",
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "pandas>=2.2.3",
    "pandasql>=0.7.3",
    "streamlit-geolocation>=0.0.10",
    "streamlit>=1.39.0",
    "torch>=2.4.1",
    "transformers>=4.45.2",
    "watchdog>=5.0.3",
]

[tool.uv]
dev-dependencies = [
    "googlemaps>=4.10.0",
    "ipykernel>=6.29.5",
    "overpy>=0.7",
    "ruff>=0.6.9",
]
